Crud de usuario finalizado

+++ src/js/acessoDados/usuarios.js
@@ -1,12 +1,12 @@
 const URL = 'http://localhost:3000'
+const URLprodutos = 'http://localhost:3000/produtos'
 
 // Metodo que retorna todos os usuarios do banco
 function getAllUsuarios() {
     fetch(URL + '/usuarios')
         .then(response => response.json())
         .then(data => {
-            // Aqui estão os dados do arquivo JSON
-            console.log(data);
+            return data;
         })
         .catch(error => {
             console.error('Erro ao ler o arquivo JSON:', error);
@@ -18,8 +18,7 @@ function getUsuarioById(id) {
     fetch(URL + '/usuarios/' + id)
         .then(response => response.json())
         .then(data => {
-            // Aqui estão os dados do arquivo JSON
-            console.log(data);
+            return data;
         })
         .catch(error => {
             console.error('Erro ao ler o arquivo JSON:', error);
@@ -28,11 +27,21 @@ function getUsuarioById(id) {
 
 // Metodo que deleta o usuário indicado pelo id
 function deleteUsuario(id){
+    let produtos = getProdutoByUserId(id);
+
+    // valida se o usuario possui produtos caso exista não deixa deletar
+    if(produtos != null && produtos != undefined && produtos.length > 0){
+        return false;
+    }
+
     fetch(URL + '/usuarios/' + id, {
         method: 'DELETE',
     })
     .then(response => response.json())
-    .then(() => location.reload());
+    .then(() => location.reload())
+    .catch(error => {
+        console.error('Erro ao acessar banco:', error);
+    });
 }
 
 // Metodo que adiciona o usuário ao banco
@@ -45,7 +54,10 @@ function createUsuario(usuario){
         body: usuario
     })
     .then(response => response.json())
-    .then(() => location.reload());
+    .then(() => location.reload())
+    .catch(error => {
+        console.error('Erro ao acessar banco:', error);
+    });
 }
 
 // Metodo que atualiza um usuário já existente
@@ -58,5 +70,20 @@ function updateUsuario(usuario){
         body: usuario
     })
     .then(response => response.json())
-    .then(() => location.reload());
+    .then(() => location.reload())
+    .catch(error => {
+        console.error('Erro ao acessar banco:', error);
+    });
+}
+
+//Metodo que busca todos os produtos que pertecem ao usuário
+function getProdutoByUserId(userId){
+    fetch(URLprodutos)
+        .then(response => response.json())
+        .then(data => {
+            return data.filter(objeto => objeto.usuarioId === userId);
+        })
+        .catch(error => {
+            console.error('Erro ao acessar banco:', error);
+        });
 }
